[REFACTOR] UserServiceTest에서 담당하던 validation 체크 테스트를 UserFactoryTest로 이동합니다
